requirements.txt
# Flask Backend
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
python-dotenv==1.0.0

# Database
pymongo==4.6.0

# Data Processing
pandas==2.3.3
numpy==1.26.2
scipy==1.11.4

# Machine Learning
scikit-learn==1.7.2
scikit-surprise==1.1.4

# NLP
nltk==3.8.1

# Visualization
matplotlib==3.8.2
seaborn==0.13.0

# Dataset
kagglehub==0.2.5

# Utils
tqdm==4.66.1
python-dateutil==2.8.2
